Increase processing speed by 5x.
This is done by ensuring that the BufReader's buffer is filled first, then processing the entire buffer in one chunk.

The way it was previously done, each byte was a fallible read, meaning the byte it failed on was observable. Because of that, the compiler was unable to further optimize the processing to make use SIMD instructions.